apartment moved to separate package
