+++ Kamienica/src/main/java/kamienica/apartment/ApartmentDaoImpl.java
@@ -1,8 +1,9 @@
-package kamienica.dao;
+package kamienica.apartment;
 
 import org.springframework.stereotype.Repository;
 
-import kamienica.model.Apartment;
+import kamienica.dao.AbstractDao;
+import kamienica.dao.DaoInterface;
 
 @Repository("apatmentDao")
 public class ApartmentDaoImpl extends AbstractDao<Integer, Apartment> implements DaoInterface<Apartment> {
