+++ Kamienica/src/main/java/kamienica/core/DivisionManager.java
@@ -6,7 +6,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
 
-import kamienica.model.Apartment;
+import kamienica.apartment.Apartment;
 import kamienica.model.Division;
 import kamienica.model.Tenant;
