+++ Kamienica/src/main/java/kamienica/dao/ReadingEnergyDAOImpl.java
@@ -6,7 +6,7 @@ import org.hibernate.Query;
 import org.hibernate.criterion.Order;
 import org.springframework.stereotype.Repository;
 
-import kamienica.model.Apartment;
+import kamienica.apartment.Apartment;
 import kamienica.model.InvoiceEnergy;
 import kamienica.model.ReadingEnergy;
