+++ Kamienica/src/main/java/kamienica/dao/ReadingWaterDAOImpl.java
@@ -8,7 +8,7 @@ import org.hibernate.Query;
 import org.hibernate.criterion.Order;
 import org.springframework.stereotype.Repository;
 
-import kamienica.model.Apartment;
+import kamienica.apartment.Apartment;
 import kamienica.model.InvoiceGas;
 import kamienica.model.InvoiceWater;
 import kamienica.model.ReadingWater;
