+++ Kamienica/src/main/java/kamienica/model/Division.java
@@ -14,6 +14,8 @@ import javax.persistence.TemporalType;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.format.annotation.DateTimeFormat;
 
+import kamienica.apartment.Apartment;
+
 @Entity
 @Table(name="division")
 public class Division {
