+++ Kamienica/src/main/java/kamienica/model/MeterAbstract.java
@@ -10,6 +10,8 @@ import javax.persistence.MappedSuperclass;
 import org.hibernate.validator.constraints.NotEmpty;
 import org.springframework.beans.factory.annotation.Autowired;
 
+import kamienica.apartment.Apartment;
+
 @MappedSuperclass
 @Inheritance
 public abstract class MeterAbstract {
