+++ Kamienica/src/main/java/kamienica/model/MeterGas.java
@@ -6,6 +6,8 @@ import javax.persistence.Table;
 
 import org.springframework.beans.factory.annotation.Autowired;
 
+import kamienica.apartment.Apartment;
+
 @Entity
 @Table(name="metergas")
 public class MeterGas extends MeterAbstract {
