+++ Kamienica/src/main/java/kamienica/model/MeterWater.java
@@ -6,6 +6,8 @@ import javax.persistence.Table;
 
 import org.springframework.beans.factory.annotation.Autowired;
 
+import kamienica.apartment.Apartment;
+
 @Entity
 @Table(name="meterwater")
 public class MeterWater extends MeterAbstract {
