+++ Kamienica/src/main/java/kamienica/model/Tenant.java
@@ -17,6 +17,8 @@ import org.hibernate.validator.constraints.NotEmpty;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.format.annotation.DateTimeFormat;
 
+import kamienica.apartment.Apartment;
+
 @Entity
 @Table(name="tenant")
 public class Tenant {
