+++ Kamienica/src/main/java/kamienica/service/ReadingServiceImpl.java
@@ -7,9 +7,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import kamienica.apartment.Apartment;
 import kamienica.dao.ReadingDao;
 import kamienica.dao.ReadingWaterDAO;
-import kamienica.model.Apartment;
 import kamienica.model.InvoiceEnergy;
 import kamienica.model.InvoiceGas;
 import kamienica.model.ReadingEnergy;
