+++ Kamienica/src/main/java/kamienica/validator/PaymentValidator.java
@@ -1,11 +1,11 @@
-package kamienica.validator;
-
-public class PaymentValidator {
-
-	boolean generalValidation;
-	boolean insertEnergy;
-	boolean insertWater;
-	boolean insertGas;
-	String errorMessage;
-	
-}
+//package kamienica.validator;
+//
+//public class PaymentValidator {
+//
+//	boolean generalValidation;
+//	boolean insertEnergy;
+//	boolean insertWater;
+//	boolean insertGas;
+//	String errorMessage;
+//	
+//}
