+++ Kamienica/src/test/java/kamienica/controller/PaymentControllerTest.java
@@ -16,8 +16,8 @@ import org.testng.Assert;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
+import kamienica.apartment.Apartment;
 import kamienica.controller.view.PaymentController;
-import kamienica.model.Apartment;
 import kamienica.model.Division;
 import kamienica.model.InvoiceEnergy;
 import kamienica.model.InvoiceGas;
