+++ Kamienica/src/test/java/kamienica/dao/ReadingWaterDaoTest.java
@@ -7,7 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
-import kamienica.model.Apartment;
+import kamienica.apartment.Apartment;
 import kamienica.model.MeterWater;
 import kamienica.model.ReadingWater;
